refactor: move engine types out (#75)
Description
---
Moves the instruction out to a common crate that can be used by wasm clients

Motivation and Context
---
I wanted to create a WASM client library that signs an instruction. These structs (specifically `hasher` and InstructionSignature) are in the engine crate. The engine crate is not suitable to be called from a WASM target, so I moved just the structs I needed to another crate. I considered moving the types to `dan_common_types`, but that didn't feel like the right place for clients, as it mainly has common hotstuff primitives
 
How Has This Been Tested?
---
Called from a separate lib, referenced and compiled to wasm (via wasm-bindgen)