feat: vn register only when registered (#105)
Description
---
This does three things:
- Ensures the vn only auto registers after being manually registered once
- Creates an epoch event channel so a caller can subscribe to epoch changes
- Retries VN registration if registration fails

Motivation and Context
---
- We don't want autoregister firing if someone hasn't manually registered
- The epoch event subscription is just handier than a poller

How Has This Been Tested?
---
Manually

~~# WIP
This has a known issue.
I was using the EpochManagerService to emit the EpochChanged event based on someone calling `EpochManagerRequest::UpdateEpoch`. Assumedly if someone was updating the epoch then the epoch has changed. This assumption was incorrect, it's not until further down the stack (`base_layer_epoch_manager`) that the block tip is actually looked at and we figure out if the epoch has infact changed.~~

Closes #62 
Closes #59