fix: use detach HSflag for tor (#132)
Description
---
Adds the DETACH hidden service flag for the tor transport on DAN layer

Motivation and Context
---
Currently we're starting and stopping the VNs a lot, each time the hidden service is recreated because it is associated with the control port session. DETACH allows the onion address to remain registered when the vn is offline. This causes a failed assertion when the VN attempts to reregister the onion, however the onion still works and the assertion is ignored.

How Has This Been Tested?
---