fix(cucumber): reduce number of runtimes, use client for get_identity (#143)
Description
---
- Remove unnecessary runtimes and extra OS threads for base node and miner processes
- Use tokio main macro in VN and cucumber main 
- Add get_identity request to JSON rpc client
- Use validator node JSON rpc client for get_identity request in cucmber
- tempfix: use different ports for cucumber tests than from regular nodes to prevent conflict

Motivation and Context
---
Tokio multiplexes tasks onto `num_cores` (multi-threaded default) worker OS-threads. It is not recommended to use more than one runtime for a process [citation needed]. The wallet functions take in a runtime, which necessitates the use of yet another OS thread to be able to create a runtime. There is no reason that I can see that the wallet should take in a runtime (it simply calls block_on and spawn internally) EDIT: looking closer, TUI needs to run on the main thread, but this wont be required for non-tui modes. TL;DR this would require a larger refactor for a small benefit. 

These are my thoughts on multiple runtimes [looking for a good explainer which debunks/validates]: 
1. each runtime spawns `num_cores` worker OS-threads
2. therefore, more stack memory is used and more tokio reactor code needs to be scheduled by the OS on the same number of cores
3. this provides no benefit at the cost of more resources

How Has This Been Tested?
---
Cucumber tests pass