fix: check pledges against changes being made (#181)
Description
---
Validate the changes made after running the WASM to make sure that all the shards have pledged it correctly.
Also fixes a bug where the DOWN and the UP of a component were not created.

Motivation and Context
---
This is a critical check that all committees are prepared to make the changes to the involved shards when the data is finally committed. 

Note that there are quite a few ugly implementation details here. The substate change map is inelegant by using a `match` instead of an `if` because of the inner data in the Substate. There are already a number of changes in this PR and I'm worried that it will cause conflicts with other PRs. I'll return to clean up that code after this and the other changes are merged

How Has This Been Tested?
---
Manually. Calling the Counter::increase method now creates a DOWN and an UP row in the sqlite db