test: use temp dir instead of manual drop (#233)
Replaced manual drop with tempdir.

When running tests on Windows, the tests would fail because the file was used by another process. Tests now pass. I assume there is some edge case that tempdir handles better.

I removed the ability to not delete a db, it wasn't used as far as I can see. If a test needs to keep a file around, it's better to use the normal store and manually delete the file afterward

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>