fix: only leader should accept/propose new vote message (issue #182) (#249)
Description
---
Definitely not the most interesting PR, but it seems that issue #182 was already been dealt with. We remove the `#TODO` reference. Our hope is that this is brought to light and checks if there was something thought to be added that was not mentioned explicitly in the issue.

Motivation and Context
---

How Has This Been Tested?
---