fix: removes pacemaker pending wait overs that already have timed out (#338)
Description
---
The PR #321 introduces a pacemaker service. As opposed to a previous implementation (see PR #307), we were not removing pending wait overs, after they timed out, which increased linearly over time the size of the collection storing these. This was not intended, and this PR fixes this problem.

Motivation and Context
---

How Has This Been Tested?
---
Manually, by printing pending_timeouts in current units and check that those that have timed out are removed.