feat: add a remove pending timeout unit test to pacemaker (#350)
Description
---
We test that pending timeouts are being removed correctly, within the internal state of a `Pacemaker` struct. Ideally, this test would make use of raw/smart pointers while `Pacemaker` run on the background (within a spawned thread) as a service. But this turns out to be complex, as we don't know which new memory address is the data pointing to. For this reason, we encapsulated most of the logic of the `run` function into proper segments, so that we can test these in a unit test environment.

Motivation and Context
---
Related to a previous bug, that we had before.

How Has This Been Tested?
---
Add a specific unit test.