feat: multiple commitments in resource, proof specify inputs (#400)
Description
---
- Confidential resources can hold multiple commitments 
- Adds the ability to join commitments
- Adds new unit test for joining multiple commitments within a vault

Motivation and Context
---
When a deposit is received from a sender bucket, output commitments are added to a vault. Since the sender has control over the mask we cannot aggregate the commitment since that would render the whole commitment unspendable. This PR re-adds the list of commitments so that a sender wallet may interrogate the commitments to determine if they are spendable. It is likely that wallets will use a mask derived from Diffie-Helman (likely [RFC-0203_StealthAddresses](https://rfc.tari.com/RFC-0203_StealthAddresses.html)).

How Has This Been Tested?
---
New unit tests