feat(cli): generate output proof for minting, support file blob args (#429)
Description
---
feat: add endpoint for generating output proofs for minting
feat: support base64 blob in cli args and manifest globals

Motivation and Context
---
When minting a confidential resource, an output commitment proof is
required. This PR adds cli tools to support this.


The key used for the proof is from the key manager but we wont know
which key to use for the transfer proof.
This is out of scope and requires PRs (#427 #425) to be merged before
this can be worked on.

We may need the walletd to associate a key for all owned vaults
regardless if they are part of an Account component so that transfer
proofs can be generated.

How Has This Been Tested?
---
Manually, `ConfidentialFaucet::mint` template call.

What process can a PR reviewer use to test or verify this change?
---
```shell
$ cargo run --bin tari_dan_wallet_cli --  proofs generate 10000 -o base64 > ./proof.txt   
$ cargo run --bin tari_dan_wallet_cli -- transactions submit-manifest -n3 ./manifest.rs -g proof=file://./proof.txt
```

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify