fix: json error on transaction page (#434)
Description
---
Fix for error on transaction page.
It works on the mock server, but please test if this fixes the error on
a live server.

Motivation and Context
---
Clicking on a recent transaction gave the following error:
"Cannot convert undefined or null to object"

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
From the homepage, click on any recent transaction.
Also click on the dropdown below "Justify" on the transaction page.

Breaking Changes
---
x

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify