refactor!: adding custom binary tags to template lib types (#446)
Description
---
* Created a `BinaryTag` enum to keep track of all binary tags
* Added `ciborium` binary tags to the relevant `template_lib` types
* Adding binary tags to `Amount` broke the _newtype_ constructor (e.g.
`Amount(1000)`) so we now need to use a explicit constructor (e.g.
`Amount::new(1000)`) in all templates and tests

Motivation and Context
---
As a result of https://github.com/tari-project/tari-dan/pull/430 now
substate data has a self-describing binary format that can be decoded
and interpreted from outside the network.

Even if all types are self-describing, we need a way to differentiate
types in the `template_lib` crate from user defined types. For that we
can use [CBOR binary
tags](https://docs.rs/ciborium/latest/ciborium/tag/index.html#) so we
can "mark" those types and identify them easily in any external
application (for example the `indexer`).

The goal of this PR is to add binary tags to the relevant types in
`template_lib` that we would want to identify in external applications:

- Amount
- ComponentAddress
- Metadata
- NonFungibleAddress
- ResourceAddress
- VaultId

How Has This Been Tested?
---
The existing unit and integration tests pass

What process can a PR reviewer use to test or verify this change?
---
This change affects all use cases. To verify the change a reviewer can:
- manually spawn a new network, deploy some templates and call methods
on them that update the state
- perform a burn claim

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Breaks at all levels, all existing data must be deleted
and VNs will not be able to sync with older versions prior to this
change

BREAKING CHANGE: Breaks at all levels, all existing data must be deleted
and VNs will not be able to sync with older versions prior to this
change