feat(tari_indexer): decode and interpret CBOR binary data (#457)
Description
---
- Inside the `tari_indexer` application:
- Created a `substate_decoder` crate that recursively decodes the CBOR
data, and is also able to interpret and extract related substate
addresses.
- Related substates of tracked substates are stored in the sqlite db.
- Refactored existing JSON RPC method response with the substate data,
because now it does not map exactly to a `Substate` type after decoding
from CBOR.
- New JSON RPC method to get all NFT collections (needed for the web
ui), that needed also a new method in the sqlite crate.
- In the `template_lib` crate added a utility function that identifies a
binary tag from an integer value, needed for the `tari_indexer` to know
how to interpret tags.
- In the `tari_indexer_web_ui` added a new section to list tracked NFT
collections. By clicking a row, a new page is shown with an image
gallery of all the NFTs of the collection. For testing purposes I used
images with _free license_ from Freepik. We can see an screenshot below:
<img width="1401" alt="Screenshot 2023-04-04 at 23 18 16"
src="https://user-images.githubusercontent.com/47919901/230023127-c6bba4ab-ded3-4dbc-b82e-3a663dad1677.png">

**TBD: paging in the NFT image gallery**. The needed JSON RPC method for
implementing it are in place and mapped inside React, ready to use.

Motivation and Context
---
Ater https://github.com/tari-project/tari-dan/pull/446 substate data now
has a self describing binary format with some lib types (addresses,
vault ids, etc.) being tagged. The next step in this process is to
update the `tari_indexer` application so it can decode that data and
make use of it.

The `tari_indexer` application can use the decoded data for two main
purposes:
- Display a gallery of all the NFTs in a collection
- Find related addresses of a tracked address, and and those to the
tracking system as well (e.g. all the vaults in an `Account` component).

How Has This Been Tested?
---
Modified the cucumber integration test related to the indexer to:
- Add an `Account` component to the tracked addresses (so we can see
that related substates are added)
- Modified the `basic_nft` template so we can specify the name and image
url as parameters when minting

What process can a PR reviewer use to test or verify this change?
---
- Uncomment the last lines of the `indexer.feature` integration test
(that show the test world and pauses the execution), run the test and
open the indexer URL in a browser.
- Manually spawn the indexer connected to an existing VN

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted (on the `tari_indexer`
application only)
- [ ] Other - Please specify

---------

Co-authored-by: stringhandler <mikethetike@tari.com>