feat!: make resource address deterministic (#450)
Description
---
Make resource addresses deterministic, specifically:
`Hash(template_address, token_symbol)`

Motivation and Context
---
Currently, the resource address is unpredictable, making some contracts
difficult. For example, you can't create a resource and mint an NFT in
the same transaction because the wallet does not know which shards to
include in it.

How Has This Been Tested?
---
Unit tests, and tests with NFT contracts

What process can a PR reviewer use to test or verify this change?
---
Create an NFT or fungible token and check the resource address created
is the same.

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Resources will be created in a new place. Existing
templates must be recompiled.

BREAKING CHANGE: Resources will be created in a new place. Existing
templates must be recompiled.