feat: adds columns to substate table for claiming fees by leader VNs (#453)
Description
---
The main goal of this PR is to add fee columns to the `substates` table.
In this way, VNs of a committee for a `(ShardId, PayloadId)` will store
the fee state of the transaction, for the respective leaders. This state
will then allow VNs to claim fees over transactions in which they were
leaders and committed new substates. This PR will also help to analyse
the possible rentability of fees to VNs.

Motivation and Context
---

How Has This Been Tested?
---
Existing tests

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify