feat: add webrtc and webextension for tari (#449)
Description
---
Adds a webrtc connectivity to wallet daemon. Currently you can start the
webrtc with jrpc. Because it needs session id.
With the jwt token it loads necessary data (SDP and ICE candidates) from
a signaling server.
Adds a webextension for tari. This extension creates a webrtc peer
connection it needs to be added via the popup. There are several
variables that the page can use accessible in window object :
- `tari_token` JWT Token
- `tari` - This is the tari object that has several methods
- `setAnswer()` - this should be called after the wallet daemon sets its
answer and ices
- `sendMessage(method, ...args)` - this calls the json rpc method on
wallet daemon with arguments if any, this returns promise which will be
resolved once the wallet daemon sends it result

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
To compile the web extension:
- cd applications/tari_web_extension
- npm run build
- in chrome go to extensions, turn on developer mode
- load unpacked from applications/tari_web_extension/pkg/

When you have the new daemon running. You can go to any page open up
console. The web extension is already injected (if you did load the page
after the web extension was installed). Get the session id
`window.tari_token`. Somehow call the jrpc jrpc method `webrtc.start`
with that token. Now call `window.tari.setAnswer()`. You should see a
message in the console that the data channel is open. Now you can call
the `tari.sendMessage`. The default callback is just logging the
responses to the console.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: stringhandler <mikethetike@tari.com>