refactor!: remove CBOR binary tag for Amount (#468)
Description
---
- remove the binary tag of the `Amount` type and refactor related tests

Motivation and Context
---
On https://github.com/tari-project/tari-dan/pull/446 we refactored the
binary schema of template types to use CBOR. To be able to discriminate
template lib types from outside the network, we added custom CBOR binary
tags to those types.

One of them was the `Amount` type, but that type should not really need
to be custom tagged and that causes more trouble that it solves. So the
purpose of this PR is just to remove the binary tag of the `Amount`
type.

How Has This Been Tested?
---
Existing unit and integration tests pass

What process can a PR reviewer use to test or verify this change?
---
This change affects all use cases. To verify the change a reviewer can
manually spawn a new network, deploy some templates and call methods on
them that update the state.

Breaking Changes
---
- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Breaks at all levels, all existing data must be deleted
and VNs will not be able to sync with older versions prior to this
change

BREAKING CHANGE: Breaks at all levels, all existing data must be deleted
and VNs will not be able to sync with older versions prior to this
change